stock-agent-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-agent-mcp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastmcp[cli]",
  "python-dotenv",
  "requests",
  "langchain-core",
  "langgraph",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["langgraph_agent*"]

[tool.mcp]
entrypoint = "server:mcp"
